+++ eth_defi/event_reader/web3factory.py
@@ -12,7 +12,7 @@ from web3 import HTTPProvider, Web3
 
 from eth_defi.chain import install_chain_middleware, install_retry_middleware, install_api_call_counter_middleware
 from eth_defi.event_reader.fast_json_rpc import patch_web3
-
+from eth_defi.provider.multi_provider import create_multi_provider_web3
 
 _web3_thread_local_cache = local()
 
@@ -55,15 +55,17 @@ class TunedWeb3Factory(Web3Factory):
 
     def __init__(
         self,
-        json_rpc_url: str,
+        rpc_config_line: str,
         http_adapter: Optional[HTTPAdapter] = None,
         thread_local_cache=False,
         api_counter=False,
     ):
         """Set up a factory.
 
-        :param json_rpc_url:
-            Node JSON-RPC server URL.
+        :param rpc_config_line:
+            JSON-RPC config line.
+
+            See :py:mod:`eth_defi.provider.multi_provider`.
 
         :param http_adapter:
             Connection pooling for HTTPS.
@@ -81,7 +83,7 @@ class TunedWeb3Factory(Web3Factory):
             Enable API counters
 
         """
-        self.json_rpc_url = json_rpc_url
+        self.rpc_config_line = rpc_config_line
 
         if not http_adapter:
             http_adapter = HTTPAdapter(pool_connections=10, pool_maxsize=10)
@@ -112,14 +114,7 @@ class TunedWeb3Factory(Web3Factory):
         session = requests.Session()
         session.mount("https://", self.http_adapter)
 
-        web3 = Web3(HTTPProvider(self.json_rpc_url, session=session))
-
-        # Enable faster ujson reads
-        patch_web3(web3)
-
-        web3.middleware_onion.clear()
-        install_chain_middleware(web3)
-        install_retry_middleware(web3)
+        web3 = create_multi_provider_web3(self.rpc_config_line, session=session)
 
         if self.thread_local_cache:
             _web3_thread_local_cache.web3 = web3